Refactor unsigned -> size_t.

+++ common/coordinate_compression.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "common/base.h"
+
 #include <algorithm>
 #include <set>
 #include <unordered_map>
@@ -10,7 +11,7 @@ template <class TValue>
 class CoordinateCompression {
  protected:
   std::vector<TValue> new_to_old;
-  std::unordered_map<TValue, unsigned> old_to_new;
+  std::unordered_map<TValue, size_t> old_to_new;
 
  public:
   CoordinateCompression() {}
@@ -21,7 +22,7 @@ class CoordinateCompression {
     new_to_old.reserve(size);
     old_to_new.clear();
     for (; begin != end; ++begin) {
-      old_to_new[*begin] = unsigned(new_to_old.size());
+      old_to_new[*begin] = new_to_old.size();
       new_to_old.push_back(*begin);
     }
   }
@@ -42,24 +43,22 @@ class CoordinateCompression {
     InitUnsorted(v);
   }
 
-  unsigned Size() const { return unsigned(new_to_old.size()); }
+  size_t Size() const { return new_to_old.size(); }
 
-  const TValue& GetOld(unsigned new_value) const {
-    return new_to_old[new_value];
-  }
+  const TValue& GetOld(size_t new_value) const { return new_to_old[new_value]; }
 
-  unsigned GetNew(const TValue& old_value) const {
+  size_t GetNew(const TValue& old_value) const {
     auto it = old_to_new.find(old_value);
     assert(it != old_to_new.end());
     return it->second;
   }
 
-  unsigned LowerBound(const TValue& old_value) const {
+  size_t LowerBound(const TValue& old_value) const {
     auto it = std::lower_bound(new_to_old.begin(), new_to_old.end(), old_value);
     return it - new_to_old.begin();
   }
 
-  unsigned UpperBound(const TValue& old_value) const {
+  size_t UpperBound(const TValue& old_value) const {
     auto it = std::upper_bound(new_to_old.begin(), new_to_old.end(), old_value);
     return it - new_to_old.begin();
   }
